+++ thrift/src/main/scala/com/twitter/finatra/thrift/internal/ThriftMethodService.scala
@@ -1,6 +1,7 @@
 package com.twitter.finatra.thrift.internal
 
-import com.twitter.finagle.{Filter, Service}
+import com.twitter.finagle
+import com.twitter.finagle.{Service, ServiceFactory, Stack}
 import com.twitter.scrooge.ThriftMethod
 import com.twitter.util.Future
 
@@ -9,13 +10,20 @@ private[thrift] class ThriftMethodService[Args, Result](
   val service: Service[Args, Result]
 ) extends Service[Args, Result] {
 
-  private[this] var filter: Filter[Args, Result, Args, Result] = Filter.identity
+  private[this] val leaf: Stack[ServiceFactory[Args, Result]] =
+    Stack.leaf(finagle.stack.Endpoint, ServiceFactory.const(service))
+
+  private[this] var stack: Stack[ServiceFactory[Args, Result]] = leaf
+
+  private[this] lazy val filteredService: Service[Args, Result] =
+    // Materialize the stack with default params.
+    Service.pending(stack.make(Stack.Params.empty)())
 
   private[finatra] def name: String = method.name
 
-  override def apply(request: Args): Future[Result] =
-    filter.andThen(service)(request)
+  private[finatra] def setStack(newStack: Stack[ServiceFactory[Args, Result]]): Unit =
+    stack = newStack ++ leaf
 
-  private[finatra] def setFilter(f: Filter.TypeAgnostic): Unit =
-    filter = filter.andThen(f.toFilter[Args, Result])
+  def apply(request: Args): Future[Result] =
+    filteredService(request)
 }